+++ config/routes.rb
@@ -1,5 +1,5 @@
 Konacha::Engine.routes.draw do
-  get '/iframe/*name' => 'specs#iframe'
-  get '/'             => 'specs#parent'
-  get '*path'         => 'specs#parent'
+  get '/iframe/*name' => 'specs#iframe', :format => false, :as => :iframe
+  root :to    => 'specs#parent'
+  get '*path' => 'specs#parent', :format => false
 end