+++ lib/konacha/engine.rb
@@ -20,19 +20,28 @@ module Konacha
       end
     end
 
-    initializer "konacha.environment" do |app|
-      unless app.config.assets.enabled
-        raise RuntimeError, "konacha requires the asset pipeline to be enabled"
+    def self.formatters
+      if ENV['FORMAT']
+        ENV['FORMAT'].split(',').map do |string|
+          eval(string).new(STDOUT)
+        end
+      else
+        [Konacha::Formatter.new(STDOUT)]
       end
+    end
 
+    initializer "konacha.environment" do |app|
       options = app.config.konacha
 
-      options.spec_dir    ||= "spec/javascripts"
-      options.port        ||= 3500
-      options.application ||= self.class.application(app)
-      options.driver      ||= :selenium
-      options.stylesheets ||= %w(application)
-      options.verbose     ||= false
+      options.spec_dir     ||= "spec/javascripts"
+      options.spec_matcher ||= /_spec\.|_test\./
+      options.port         ||= 3500
+      options.application  ||= self.class.application(app)
+      options.driver       ||= :selenium
+      options.stylesheets  ||= %w(application)
+      options.verbose      ||= false
+      options.runner_port  ||= nil
+      options.formatters   ||= self.class.formatters
 
       app.config.assets.paths << app.root.join(options.spec_dir).to_s
     end