+++ spec/spec_helper.rb
@@ -13,11 +13,19 @@ require "ejs"
 
 require "capybara/rails"
 require "capybara/firebug"
+require "capybara/poltergeist"
 
 Capybara.configure do |config|
   config.default_selector = :css
   config.default_driver   = :selenium_with_firebug
   config.app              = Konacha.application
+end
+
+Capybara.register_driver :poltergeist do |app|
+  # Work around a bug in PhantomJS where `return true` from a
+  # window.onerror handler does not prevent an uncaught exception
+  # from being reported to Ruby.
+  Capybara::Poltergeist::Driver.new(app, :js_errors => false)
 end
 
 module Konacha